requirements.txt
langchain
langgraph
langchain-google-genai
langchain-community
faiss-cpu
python-dotenv